Moved the `ASCII` class from `TopCommand.cs` to a new file `ASCII.cs`.
Additionally, updated a comment in `TopCommand.cs` from German
to English for better accessibility and clarity.
